pyproject.toml
[project]
name = "bitpin-task"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "psycopg2-binary>=2.9.10",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
